Made small changes. started inverseCDF.
