+++ rbig/__init__.py
@@ -1,0 +1,2 @@
+from .rbig import RBIG, RBIGMI, RBIGKLD
+from .univariate import HistogramUnivariateDensity
